Implemented basic combat and refactored follower and building classes
Soldiers can now be commanded to attack enemies, which can be hit when in range. The enemy will then start to chase the soldier that attacked it.

Also majorly refactored building and follower classes to instead be base classes for use in 'sub-classes' (i.e. Soldiers, Workers for Follower or Storage, Wall for Building). This allows for new building or follower variants to be more easily set up.